+++ sbomify/licensing/__init__.py
@@ -1,0 +1,5 @@
+from .loader import ALL_LICENSES, load_licenses, validate_expression
+
+load_licenses()
+
+__all__ = ["ALL_LICENSES", "load_licenses", "validate_expression"]